addr review comments, move some tests
